requirements.txt
streamlit==0.82.0
pandas==1.4.2
numpy==1.22.3
matplotlib==3.5.2
Pillow==9.1.0
